add: player and scoreBoard(not finished)

+++ scripts/init.sql
@@ -4,6 +4,6 @@ CREATE TABLE IF NOT EXISTS ranking (
 
     name TEXT NOT NULL,
     apples INTEGER NOT NULL CHECK (apples >= 0),
-    time REAL NOT NULL CHECK (time >= 0),
+    time INTEGER NOT NULL CHECK (time >= 0),
     score REAL NOT NULL CHECK (score >= 0)
 );
